chore(signup) put signup validation in middleware [FINISHES #167216160]
