+++ src/middleware/strict_validate.js
@@ -1,0 +1,9 @@
+import { strictValidate } from '../helper/user';
+import { results, ERROR } from '../helper/result';
+
+export default (req, res, next) => {
+    const error = strictValidate(req);
+    if (error) return res.status(400).send(results(400, ERROR, error));
+
+    next();
+}
